refatorando o projeto, usando DTOs e conectando ao banco

+++ src/main/java/com/teste/restful/service/ProdutoService.java
@@ -2,18 +2,22 @@ package com.teste.restful.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.teste.restful.model.ProdutoModel;
-import com.teste.restful.repository.ProdutoRepository;
+import com.teste.restful.model.Produto;
+import com.teste.restful.model.exception.ResourceNotFoundException;
+import com.teste.restful.repository.ProdutoRepositoryOLD;
+import com.teste.restful.shared.ProdutoDTO;
 
 @Service
 public class ProdutoService {
 
 	@Autowired
-	private ProdutoRepository produtoRepository;
+	private ProdutoRepositoryOLD produtoRepository;
 	
 	/**
 	 * Método para retornar todos os produtos
@@ -21,8 +25,12 @@ public class ProdutoService {
 	 * @param
 	 * @return List<ProdutoModel> - lista de produtos
 	 */
-	public List<ProdutoModel> findAll() {
-	    return produtoRepository.findAll();
+	public List<ProdutoDTO> findAll() {
+		List<Produto> produtos = produtoRepository.findAll();
+		return produtos
+				.stream()
+				.map(produto -> new ModelMapper().map(produto, ProdutoDTO.class))
+				.collect(Collectors.toList());
     }
 	
 	
@@ -32,8 +40,13 @@ public class ProdutoService {
 	 * @param id - id do produto
 	 * @return ProdutoModel - produto
 	 */
-	public Optional<ProdutoModel> findById(Integer id) {
-		return produtoRepository.findById(id);
+	public Optional<ProdutoDTO> findById(Integer id) {
+		Optional<Produto> produto = produtoRepository.findById(id);
+		if(produto.isEmpty()) {
+			throw new ResourceNotFoundException("Produto não encontrado");
+		}
+		ProdutoDTO produtoDTO = new ModelMapper().map(produto.get(), ProdutoDTO.class);
+		return Optional.of(produtoDTO);
 	}
 	
 	
@@ -43,8 +56,15 @@ public class ProdutoService {
 	 * @param produto - produto a ser salvo
 	 * @return ProdutoModel - produto salvo
 	 */
-	public ProdutoModel save(ProdutoModel produto) {
-		return produtoRepository.save(produto);
+	public ProdutoDTO save(ProdutoDTO produtoDto) {
+		produtoDto.setId(null);
+		
+		ModelMapper modelMapper = new ModelMapper();
+		Produto produto = modelMapper.map(produtoDto, Produto.class);
+		produto = produtoRepository.save(produto);
+		produtoDto.setId(produto.getId());
+		return produtoDto;
+		
     }
 
 	/**
@@ -53,10 +73,12 @@ public class ProdutoService {
 	 * @param produto - produto a ser atualizado
 	 * @return ProdutoModel - produto atualizado
 	 */
-	public ProdutoModel update(Integer id, ProdutoModel produto) {
-		produto.setId(id);
-		
-        return produtoRepository.update(produto);
+	public ProdutoDTO update(Integer id, ProdutoDTO produtoDto) {
+		produtoDto.setId(id);
+		ModelMapper modelMapper = new ModelMapper();
+		Produto produtoToUpdate = modelMapper.map(produtoDto, Produto.class);
+		produtoRepository.save(produtoToUpdate);
+		return produtoDto;
 	}
 	
 	/**
@@ -66,7 +88,12 @@ public class ProdutoService {
 	 * @return
 	 */
 	public void delete(Integer id) {
+		Optional<Produto> produto = produtoRepository.findById(id);
+		
+		if(produto.isEmpty()) {
+			throw new ResourceNotFoundException("Produto "+id+" não encontrado");
+		}
+		
 		produtoRepository.deleteById(id);
-
 	}
 }
